requirements.txt
fastapi
uvicorn
python-dotenv
pydantic
langchain
langchain-google-genai
langchain-core
httpx
beautifulsoup4
